Removing deprecated code for the move to sharp as resizing engine
