+++ src/streams/sources/local.js
@@ -41,7 +41,11 @@ Local.prototype._read = function(){
 
     // if there is an error store it on the image object and pass it along
     if (err) {
-      _this.image.error = new Error(err);
+      _this.image.error = err;
+
+      if (err.code === 'ENOENT') {
+        _this.image.error.statusCode = 404;
+      }
     }
 
     // if not store the image buffer
