+++ src/streams/sources/twitter.js
@@ -92,11 +92,9 @@ Twitter.prototype._read = function(){
         .replace('_mini', '');
 
       imgStream = request.get(imageUrl);
-      imgStream.on('data', function(d){ bufs.push(d); });
-      imgStream.on('error', function(err){
-        _this.image.error = new Error(err);
-      });
-      imgStream.on('end', function(){
+      imgStream.on('data',  function(d) { bufs.push(d); });
+      imgStream.on('error', function(err) { _this.image.error = err; });
+      imgStream.on('end',   function() {
         _this.image.log.timeEnd('twitter');
         _this.image.contents = Buffer.concat(bufs);
         _this.image.originalContentLength = contentLength(bufs);
