--- test/index2.html
@@ -1,117 +1,0 @@
-<!doctype html>
-<html lang="en">
-<head>
-  <meta charset="UTF-8">
-  <title>Image Resizer Test Page</title>
-
-  <style>
-    body {
-      font-family: 'Open Sans', Helvetica;
-    }
-    h2 {
-      clear: both;
-    }
-    .sample-img {
-      float: left;
-      margin: 10px;
-    }
-    .sample-img.newline {
-      clear:left;
-    }
-  </style>
-
-</head>
-<body>
-
-  <h1>Image Lib Comparison: sharp vs gm</h1>
-
-  <h2>Original</h2>
-  <div class="sample-img">
-    <img src="/elocal/test/sample_images/image1.jpg?sharp" style="height:200px">
-    <p>sharp: original</p>
-  </div>
-  <div class="sample-img">
-    <img src="/elocal/test/sample_images/image1.jpg" style="height:200px">
-    <p>gm: original</p>
-  </div>
-
-  <h2>Square</h2>
-  <div class="sample-img">
-    <img src="/s200-elocal/test/sample_images/image1.jpg?sharp">
-    <p>sharp: s200</p>
-  </div>
-  <div class="sample-img">
-    <img src="/s200-elocal/test/sample_images/image1.jpg">
-    <p>gm: s200</p>
-  </div>
-
-  <h2>Gravity</h2>
-  <div class="sample-img">
-    <img src="/s100-gsw-elocal/test/sample_images/image1.jpg?sharp">
-    <p>sharp: s100-gsw</p>
-  </div>
-  <div class="sample-img">
-    <img src="/s100-gsw-elocal/test/sample_images/image1.jpg">
-    <p>gm: s100-gsw</p>
-  </div>
-
-  <div class="sample-img newline">
-    <img src="/s100-gne-elocal/test/sample_images/image1.jpg?sharp">
-    <p>sharp: s100-gne</p>
-  </div>
-  <div class="sample-img">
-    <img src="/s100-gne-elocal/test/sample_images/image1.jpg">
-    <p>gm: s100-gne</p>
-  </div>
-
-
-  <h2>Crop</h2>
-  <div class="sample-img">
-    <img src="/w200-cfit-elocal/test/sample_images/image2.jpg?sharp">
-    <p>sharp: w200-cfit</p>
-  </div>
-  <div class="sample-img">
-    <img src="/w200-cfit-elocal/test/sample_images/image2.jpg">
-    <p>gm: w200-cfit</p>
-  </div>
-
-  <div class="sample-img newline">
-    <img src="/w200-cfill-elocal/test/sample_images/image2.jpg?sharp">
-    <p>sharp: w200-cfill</p>
-  </div>
-  <div class="sample-img">
-    <img src="/w200-cfill-elocal/test/sample_images/image2.jpg">
-    <p>gm: w200-cfill</p>
-  </div>
-
-  <div class="sample-img newline">
-    <img src="/w200-ccut-gne-elocal/test/sample_images/image2.jpg?sharp">
-    <p>sharp: w200-ccut-gne</p>
-  </div>
-  <div class="sample-img">
-    <img src="/w200-ccut-gne-elocal/test/sample_images/image2.jpg">
-    <p>gm: w200-ccut-gne</p>
-  </div>
-
-
-  <h2>Non-Square Crop Fill</h2>
-  <div class="sample-img">
-    <img src="/w200-h150-cfill-elocal/test/sample_images/image1.jpg?sharp">
-    <p>sharp: w200-h150-cfill</p>
-  </div>
-  <div class="sample-img">
-    <img src="/w200-h150-cfill-elocal/test/sample_images/image1.jpg">
-    <p>gm: w200-h150-cfill</p>
-  </div>
-
-  <div class="sample-img newline">
-    <img src="/w150-h200-cfill-elocal/test/sample_images/image1.jpg?sharp">
-    <p>sharp: w150-h200-cfill</p>
-  </div>
-  <div class="sample-img">
-    <img src="/w150-h200-cfill-elocal/test/sample_images/image1.jpg">
-    <p>gm: w150-h200-cfill</p>
-  </div>
-
-</body>
-</html>
